Move Cardano node API types to a different module
