+++ ouroboros-network/src/Cardano/Node/PeerSelection/Types.hs
@@ -1,0 +1,6 @@
+module Cardano.Node.PeerSelection.Types where
+
+data ChurnMode = ChurnModeBulkSync
+               | ChurnModeNormal
+               deriving Show
+
