+++ ouroboros-network/src/Ouroboros/Network/Diffusion/Common.hs
@@ -24,6 +24,7 @@ import Control.Tracer (Tracer, nullTracer)
 
 import Network.Mux qualified as Mx
 
+import Cardano.Node.PeerSelection.LocalRootPeers (OutboundConnectionsState)
 import Ouroboros.Network.Mux (OuroborosApplicationWithMinimalCtx,
            OuroborosBundleWithExpandedCtx)
 import Ouroboros.Network.NodeToClient (Versions)
@@ -34,7 +35,6 @@ import Ouroboros.Network.NodeToNode qualified as NodeToNode
 import Ouroboros.Network.PeerSelection.Governor.Types (PublicPeerSelectionState)
 import Ouroboros.Network.PeerSelection.LedgerPeers.Type
            (LedgerPeersConsensusInterface)
-import Ouroboros.Network.PeerSelection.LocalRootPeers (OutboundConnectionsState)
 import Ouroboros.Network.Snocket (FileDescriptor)
 import Ouroboros.Network.Socket (SystemdSocketTracer)
