+++ ouroboros-network/src/Ouroboros/Network/Diffusion/Configuration.hs
@@ -44,10 +44,11 @@ module Ouroboros.Network.Diffusion.Configuration
 import Control.Monad.Class.MonadTime.SI
 import System.Random (randomRIO)
 
+import Cardano.Node.ConsensusMode
+import Cardano.Node.Types (MinBigLedgerPeersForTrustedState (..))
 import Ouroboros.Network.BlockFetch (BlockFetchConfiguration (..))
 import Ouroboros.Network.ConnectionManager.Core (defaultProtocolIdleTimeout,
            defaultResetTimeout, defaultTimeWaitTimeout)
-import Ouroboros.Network.ConsensusMode
 import Ouroboros.Network.Diffusion (P2P (..))
 import Ouroboros.Network.Diffusion.Policies (closeConnectionTimeout,
            deactivateTimeout, maxChainSyncTimeout, minChainSyncTimeout,
@@ -56,8 +57,6 @@ import Ouroboros.Network.NodeToNode (DiffusionMode (..),
            MiniProtocolParameters (..), defaultMiniProtocolParameters)
 import Ouroboros.Network.PeerSelection.Governor.Types
            (ConsensusModePeerTargets (..), PeerSelectionTargets (..))
-import Ouroboros.Network.PeerSelection.LedgerPeers.Type
-           (MinBigLedgerPeersForTrustedState (..))
 import Ouroboros.Network.PeerSelection.PeerSharing (PeerSharing (..))
 import Ouroboros.Network.PeerSharing (ps_POLICY_PEER_SHARE_MAX_PEERS,
            ps_POLICY_PEER_SHARE_STICKY_TIME)
