+++ ouroboros-network/src/Ouroboros/Network/Diffusion/P2P.hs
@@ -100,12 +100,11 @@ import Ouroboros.Network.NodeToNode (AcceptedConnectionsLimit (..),
            DiffusionMode (..), NodeToNodeVersion (..),
            NodeToNodeVersionData (..), RemoteAddress)
 import Ouroboros.Network.NodeToNode qualified as NodeToNode
-import Ouroboros.Network.PeerSelection.Bootstrap (UseBootstrapPeers)
 import Ouroboros.Network.PeerSelection.Churn (PeerChurnArgs (..))
 import Ouroboros.Network.PeerSelection.Governor qualified as Governor
 import Ouroboros.Network.PeerSelection.Governor.Types
-           (ChurnMode (ChurnModeNormal), ConsensusModePeerTargets (..),
-           DebugPeerSelection (..), PeerSelectionActions, PeerSelectionCounters,
+           (ConsensusModePeerTargets (..), DebugPeerSelection (..),
+           PeerSelectionActions, PeerSelectionCounters,
            PeerSelectionInterfaces (..), PeerSelectionPolicy (..),
            PeerSelectionState, TracePeerSelection (..),
            emptyPeerSelectionCounters, emptyPeerSelectionState)
@@ -117,8 +116,7 @@ import Ouroboros.Network.PeerSelection.LedgerPeers (TraceLedgerPeers,
            WithLedgerPeersArgs (..))
 #ifdef POSIX
 import Ouroboros.Network.PeerSelection.LedgerPeers.Type (LedgerPeerSnapshot,
-           LedgerPeersConsensusInterface (..), MinBigLedgerPeersForTrustedState,
-           UseLedgerPeers)
+           LedgerPeersConsensusInterface (..), UseLedgerPeers)
 import Ouroboros.Network.PeerSelection.PeerMetric (PeerMetrics,
            fetchynessBlocks, upstreamyness)
 #else
@@ -126,7 +124,10 @@ import Ouroboros.Network.PeerSelection.LedgerPeers.Type (LedgerPeerSnapshot,
            MinBigLedgerPeersForTrustedState, UseLedgerPeers)
 import Ouroboros.Network.PeerSelection.PeerMetric (PeerMetrics)
 #endif
-import Ouroboros.Network.ConsensusMode
+import Cardano.Node.ConsensusMode
+import Cardano.Node.PeerSelection.Bootstrap (UseBootstrapPeers)
+import Cardano.Node.PeerSelection.Types (ChurnMode (..))
+import Cardano.Node.Types (MinBigLedgerPeersForTrustedState)
 import Ouroboros.Network.PeerSelection.PeerSelectionActions
 import Ouroboros.Network.PeerSelection.PeerSharing (PeerSharing (..))
 import Ouroboros.Network.PeerSelection.PeerStateActions (PeerConnectionHandle,
