+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/Governor/Types.hs
@@ -32,7 +32,6 @@ module Ouroboros.Network.PeerSelection.Governor.Types
   , PeerStateActions (..)
   , PeerSelectionActions (..)
   , PeerSelectionInterfaces (..)
-  , ChurnMode (..)
     -- * P2P governor internals
   , PeerSelectionState (..)
   , emptyPeerSelectionState
@@ -142,11 +141,9 @@ import Control.Monad.Class.MonadTime.SI
 import System.Random (StdGen)
 
 import Control.Concurrent.Class.MonadSTM.Strict
-import Ouroboros.Network.ConsensusMode
 import Ouroboros.Network.ExitPolicy
-import Ouroboros.Network.PeerSelection.Bootstrap (UseBootstrapPeers (..))
+import Cardano.Node.PeerSelection.Bootstrap (UseBootstrapPeers (..))
 import Ouroboros.Network.PeerSelection.LedgerPeers.Type
-import Ouroboros.Network.PeerSelection.LocalRootPeers (OutboundConnectionsState)
 import Ouroboros.Network.PeerSelection.PeerSharing (PeerSharing)
 import Ouroboros.Network.PeerSelection.PublicRootPeers (PublicRootPeers)
 import Ouroboros.Network.PeerSelection.PublicRootPeers qualified as PublicRootPeers
@@ -161,7 +158,10 @@ import Ouroboros.Network.PeerSelection.Types (PeerSource (..),
            PeerStatus (PeerHot, PeerWarm))
 import Ouroboros.Network.Protocol.PeerSharing.Type (PeerSharingAmount,
            PeerSharingResult (..))
-
+import Cardano.Node.Types (LedgerStateJudgement (..), MinBigLedgerPeersForTrustedState)
+import Cardano.Node.PeerSelection.Types (ChurnMode)
+import Cardano.Node.PeerSelection.LocalRootPeers (OutboundConnectionsState)
+import Cardano.Node.ConsensusMode (ConsensusMode)
 
 -- | A peer pick policy is an action that picks a subset of elements from a
 -- map of peers.
@@ -1767,7 +1767,3 @@ data DebugPeerSelection peeraddr where
 
 deriving instance (Ord peeraddr, Show peeraddr)
                => Show (DebugPeerSelection peeraddr)
-
-data ChurnMode = ChurnModeBulkSync
-               | ChurnModeNormal deriving Show
-
