+++ ouroboros-network/src/Ouroboros/Network/PeerSelection/State/LocalRootPeers.hs
@@ -38,8 +38,8 @@ import Data.Map.Strict qualified as Map
 import Data.Set (Set)
 import Data.Set qualified as Set
 
+import Cardano.Node.PeerSelection.PeerTrustable (PeerTrustable (..))
 import Ouroboros.Network.PeerSelection.PeerAdvertise (PeerAdvertise)
-import Ouroboros.Network.PeerSelection.PeerTrustable (PeerTrustable (..))
 
 
 ---------------------------------------
